refactor: reshape adapters/ -- promote schema, dissolve catalog, home log_config
The adapters/ layer mixed two directions and carried several single-purpose
sub-packages. Five changes give it one clear identity and remove the ceremony.

1. PROMOTE schema/ to top-level delta_engine/schema/.
   schema/ is the user-facing, pyspark-free DSL for DEFINING tables -- the
   driving input, not a driven technology binding. Sitting under adapters/ beside
   the Spark adapter conflated the two directions. After the move, adapters/ means
   exactly one thing (outgoing tech bindings), and the package reads as four clear
   layers: domain/ (pure logic), application/ (use cases + ports), schema/ (user
   DSL), adapters/ (technology). A second backend now lands as adapters/<backend>/
   beside databricks/, with schema/ visibly shared and backend-independent.

2. DISSOLVE adapters/databricks/catalog/.
   Its __init__ exported nothing (full-path imports only) and "catalog"
   mis-described executor.py (it runs SQL, not catalog ops). reader.py and
   executor.py move up to the databricks/ root next to build_engine.py and sql/.

3. MOVE log_config.py into adapters/databricks/.
   Its only caller is the Databricks entry point and it contains py4j/Spark-
   specific logger quieting; at the package root it falsely implied generality.
   Stdlib-only, so the pyspark-free guarantee is unaffected; configure_logging
   stays lazily exported from the package root.

4. COLLAPSE schema/delta/ into schema/.
   The delta/ nesting named a format distinction that does not exist (Delta is the
   only format; the class is already DeltaTable). table.py and properties.py move
   up. If a second format ever lands, re-nesting is one mechanical commit then.

5. DELETE the schema/types.py singletons (INTEGER/STRING/...).
   Zero internal callers; Integer() == Integer() already, so they added no value
   over the classes. schema/__init__ now imports the type CLASSES directly from
   domain.model (it previously relayed them through types.py), and types.py is
   gone.

Pure structural move: every change is a file relocation + import-path rewrite,
no logic touched. git tracked the moves as renames (history preserved). Verified:
the pyspark-free import guarantee still holds (import delta_engine with pyspark
blocked), lint + mypy clean, full unit suite passes (145; the single deleted test
covered the removed singletons). README and the review doc updated.
