+++ src/delta_engine/adapters/databricks/__init__.py
@@ -1,4 +1,4 @@
 from delta_engine.adapters.databricks.build_engine import build_databricks_engine
-from delta_engine.log_config import configure_logging
+from delta_engine.adapters.databricks.log_config import configure_logging
 
 __all__ = ["build_databricks_engine", "configure_logging"]
