--- src/delta_engine/adapters/databricks/catalog/__init__.py
@@ -1,1 +1,0 @@
-
