+++ src/delta_engine/schema/properties.py
