+++ tests/test_public_api.py
@@ -43,8 +43,8 @@ def test_eager_names_are_importable_and_identical_to_their_source():
     # Given the curated root namespace
     # Then every pyspark-free name resolves to the same object as its source module
     from delta_engine import Column, DeltaTable, Engine, Registry
-    from delta_engine.adapters.schema import Column as ColumnImpl, DeltaTable as DeltaTableImpl
     from delta_engine.application import Engine as EngineImpl, Registry as RegistryImpl
+    from delta_engine.schema import Column as ColumnImpl, DeltaTable as DeltaTableImpl
 
     assert DeltaTable is DeltaTableImpl
     assert Column is ColumnImpl
